requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
langchain
sentence-transformers
faiss-cpu
transformers
torch
numpy
pandas
scikit-learn